Remove some duplication using DefaultSignatures and type equality
Basically, we should get
```
    MonadEquiv n, MonadTrans t => MonadEquiv (t n)
```
generically by defining it once for all monad transformers `t`.
GHC offers us DefaultSignatures to this end, but a flaw in the
ointment is still that a class can only have one default
implementation per method.  Thus we need to retain some boilerplate
for `combine` and `equate`, which already have a default implementation.